[IndeterminateBar]: moved in uui
