+++ epam-assets/theme/theme_promo.scss
@@ -705,6 +705,11 @@
     --uui-file-card-dropstart-background: var(--uui-primary-lightest);
     --uui-file-card-link-hover: var(--uui-primary-dark);
 
+    /* IndeterminateBar */
+
+    --uui-indeterminate-bar-wrapper: var(--uui-secondary-light);
+    --uui-indeterminate-bar-background: var(--uui-primary);
+
     /* ------------------------------------ */
 
     .tooltip-contrast,
