--- epam-promo/components/widgets/IndeterminateBar.module.scss
@@ -1,38 +1,0 @@
-@use '../../assets/styles/colors' as *;
-
-.root {
-    position: relative;
-    background-color: $gray30;
-
-    .bar {
-        background-color: $blue;
-        width: 33%;
-        top: 0;
-        bottom: 0;
-        position: absolute;
-        animation: progressBar-indeterminate 2s linear 0.15s infinite;
-    }
-}
-
-
-@keyframes progressBar-indeterminate {
-    0% {
-        left: -33%;
-    }
-
-    100% {
-        left: 100%;
-    }
-}
-
-.size-12 {
-    height: 12px;
-}
-
-.size-18 {
-    height: 18px;
-}
-
-.size-24 {
-    height: 24px;
-}
