+++ epam-promo/components/widgets/index.ts
@@ -1,4 +1,3 @@
 export * from './Badge';
 export * from './ProgressBar';
-export * from './IndeterminateBar';
 export * from './IndicatorBar';
