+++ loveship/components/widgets/IndicatorBar.tsx
@@ -1,7 +1,8 @@
 import * as React from 'react';
 import cx from 'classnames';
 import { IHasCX, IHasRawProps } from '@epam/uui-core';
-import { ProgressBar, IndeterminateBar } from './';
+import { ProgressBar } from './ProgressBar';
+import { IndeterminateBar } from '@epam/uui';
 import css from './IndicatorBar.module.scss';
 
 interface IIndicatorProps extends IHasCX, IHasRawProps<HTMLDivElement> {
