+++ loveship/components/widgets/index.ts
@@ -3,4 +3,3 @@ export * from './Badge';
 export * from './Tag';
 export * from './ProgressBar';
 export * from './IndicatorBar';
-export * from './IndeterminateBar';
