+++ uui/components/widgets/IndeterminateBar.tsx
