+++ uui/components/widgets/__tests__/IndeterminateBar.test.tsx
