+++ uui/components/widgets/index.ts
@@ -4,3 +4,4 @@ export * from './Badge';
 export * from './Tag';
 export * from './Spinner';
 export * from './Paginator';
+export * from './IndeterminateBar';
